requirements.txt
streamlit
plotly
pandas
datetime
crewai
crewai-tools
langchain-openai
python-dotenv
sqlitecloud